moved AuthData hook to the hooks folder
