+++ src/pages/Login/Login.jsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { Button } from "@/components/ui/button.jsx";
 import { Input } from "@/components/ui/input.jsx";
-import { AuthData } from "@/components/auth/AuthWrapper.jsx";
+import { AuthData } from "@/hooks/AuthData";
 
 const Login = () => {
   const [email, setEmail] = useState("");
