+++ src/pages/Profile/UserCharts.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react'
 import { getWeightHistory, updateWeight } from "@/lib/profile";
 import LineChartComponent from "./LineChart";
-import { AuthData } from '@/components/auth/AuthWrapper';
+import { AuthData } from "@/hooks/AuthData";
 import Container from '@/components/Container';
 function GridItem({ children, className}) {
